+++ src/euroeval/dataset_configs/finnish.py
@@ -109,26 +109,6 @@ INCLUDE_FI_CONFIG = DatasetConfig(
     languages=[FINNISH],
 )
 
-WMT24PP_EN_FI_CONFIG = TranslationDatasetConfig(
-    name="wmt24pp-en-fi",
-    pretty_name="WMT24++-en-fi",
-    source="EuroEval/wmt24pp-en-fi",
-    task=TRANSLATION,
-    languages=[FINNISH],
-    source_language=ENGLISH,
-    target_language=FINNISH,
-)
-
-WMT24PP_FI_EN_CONFIG = TranslationDatasetConfig(
-    name="wmt24pp-fi-en",
-    pretty_name="WMT24++-fi-en",
-    source="EuroEval/wmt24pp-fi-en",
-    task=TRANSLATION,
-    languages=[FINNISH],
-    source_language=FINNISH,
-    target_language=ENGLISH,
-)
-
 
 # Unofficial datasets ###