+++ src/euroeval/dataset_configs/french.py
@@ -118,26 +118,6 @@ MULTILOKO_FR_CONFIG = DatasetConfig(
     val_split=None,
 )
 
-WMT24PP_EN_FR_CONFIG = TranslationDatasetConfig(
-    name="wmt24pp-en-fr",
-    pretty_name="WMT24++-en-fr",
-    source="EuroEval/wmt24pp-en-fr",
-    task=TRANSLATION,
-    languages=[FRENCH],
-    source_language=ENGLISH,
-    target_language=FRENCH,
-)
-
-WMT24PP_FR_EN_CONFIG = TranslationDatasetConfig(
-    name="wmt24pp-fr-en",
-    pretty_name="WMT24++-fr-en",
-    source="EuroEval/wmt24pp-fr-en",
-    task=TRANSLATION,
-    languages=[FRENCH],
-    source_language=FRENCH,
-    target_language=ENGLISH,
-)
-
 
 # Unofficial datasets ###