+++ src/euroeval/dataset_configs/hungarian.py
@@ -94,26 +94,6 @@ INCLUDE_HU_CONFIG = DatasetConfig(
     languages=[HUNGARIAN],
 )
 
-WMT24PP_EN_HU_CONFIG = TranslationDatasetConfig(
-    name="wmt24pp-en-hu",
-    pretty_name="WMT24++-en-hu",
-    source="EuroEval/wmt24pp-en-hu",
-    task=TRANSLATION,
-    languages=[HUNGARIAN],
-    source_language=ENGLISH,
-    target_language=HUNGARIAN,
-)
-
-WMT24PP_HU_EN_CONFIG = TranslationDatasetConfig(
-    name="wmt24pp-hu-en",
-    pretty_name="WMT24++-hu-en",
-    source="EuroEval/wmt24pp-hu-en",
-    task=TRANSLATION,
-    languages=[HUNGARIAN],
-    source_language=HUNGARIAN,
-    target_language=ENGLISH,
-)
-
 
 # Unofficial datasets ###