+++ src/euroeval/dataset_configs/lithuanian.py
@@ -101,26 +101,6 @@ INCLUDE_LT_CONFIG = DatasetConfig(
     languages=[LITHUANIAN],
 )
 
-WMT24PP_EN_LT_CONFIG = TranslationDatasetConfig(
-    name="wmt24pp-en-lt",
-    pretty_name="WMT24++-en-lt",
-    source="EuroEval/wmt24pp-en-lt",
-    task=TRANSLATION,
-    languages=[LITHUANIAN],
-    source_language=ENGLISH,
-    target_language=LITHUANIAN,
-)
-
-WMT24PP_LT_EN_CONFIG = TranslationDatasetConfig(
-    name="wmt24pp-lt-en",
-    pretty_name="WMT24++-lt-en",
-    source="EuroEval/wmt24pp-lt-en",
-    task=TRANSLATION,
-    languages=[LITHUANIAN],
-    source_language=LITHUANIAN,
-    target_language=ENGLISH,
-)
-
 
 # Unofficial datasets ###