+++ src/euroeval/dataset_configs/luxembourgish.py
@@ -1,7 +1,9 @@
 """All Luxembourgish dataset configurations used in EuroEval."""
 
-from ..data_models import DatasetConfig, TranslationDatasetConfig
-from ..languages import ENGLISH, LUXEMBOURGISH
+from euroeval.enums import ModelType
+
+from ..data_models import DatasetConfig
+from ..languages import LUXEMBOURGISH
 from ..tasks import (
     HALLU,
     INSTRUCTION_FOLLOWING,
@@ -11,7 +13,6 @@ from ..tasks import (
     RC,
     SENT,
     TEXT_CLASSIFICATION,
-    TRANSLATION,
 )
 
 # Official datasets ###
@@ -91,6 +92,7 @@ LTZGLUE_ID_CONFIG = DatasetConfig(
     task=TEXT_CLASSIFICATION,
     languages=[LUXEMBOURGISH],
     train_split=None,
+    allowed_model_types=[ModelType.GENERATIVE],
     num_few_shot_examples=0,
     labels=[
         "addtoplaylist",
@@ -119,26 +121,6 @@ MULTI_IFEVAL_LB_CONFIG = DatasetConfig(
     languages=[LUXEMBOURGISH],
     train_split=None,
     val_split=None,
-)
-
-FLORES_EN_LB_CONFIG = TranslationDatasetConfig(
-    name="flores-en-lb",
-    pretty_name="FLORES-en-lb",
-    source="EuroEval/flores-en-lb",
-    task=TRANSLATION,
-    languages=[LUXEMBOURGISH],
-    source_language=ENGLISH,
-    target_language=LUXEMBOURGISH,
-)
-
-FLORES_LB_EN_CONFIG = TranslationDatasetConfig(
-    name="flores-lb-en",
-    pretty_name="FLORES-lb-en",
-    source="EuroEval/flores-lb-en",
-    task=TRANSLATION,
-    languages=[LUXEMBOURGISH],
-    source_language=LUXEMBOURGISH,
-    target_language=ENGLISH,
 )